Reorganize generic events as a feature
