--- fe1-web/components/eventList/events/index.ts
@@ -1,3 +1,0 @@
-import Event from './Event';
-
-export default Event;
