+++ fe1-web/features/events/components/index.ts
@@ -1,0 +1,3 @@
+import Event from 'features/events/components/Event';
+
+export default Event;
