+++ fe1-web/features/events/objects/LaoEventBuilder.ts
@@ -1,6 +1,7 @@
 import { Meeting, MeetingState } from 'features/meeting/objects';
 import { Election, ElectionState } from 'features/evoting/objects';
 import { RollCall, RollCallState } from 'features/rollCall/objects';
+
 import { LaoEvent, LaoEventState, LaoEventType } from './LaoEvent';
 
 export function eventFromState(evtState: LaoEventState): LaoEvent | undefined {
