+++ fe1-web/features/events/objects/index.ts
@@ -1,0 +1,2 @@
+export * from './LaoEvent';
+export * from './LaoEventBuilder';
