+++ fe1-web/features/events/store/index.ts
@@ -1,0 +1,1 @@
+export * from './EventStore';
