+++ fe1-web/features/events/styles/eventViewStyles.ts
@@ -1,4 +1,5 @@
 import { StyleSheet } from 'react-native';
+
 import { Spacing } from 'styles';
 
 const eventViewStyles = StyleSheet.create({
