+++ fe1-web/model/objects/index.ts
@@ -8,8 +8,6 @@ export * from './ConnectToLao';
 export * from './Hash';
 export * from './KeyPair';
 export * from './Lao';
-export * from './LaoEvent';
-export * from './LaoEventBuilder';
 export * from './PopToken';
 export * from './PrivateKey';
 export * from './PublicKey';
