+++ fe1-web/model/objects/wallet/Wallet.ts
@@ -1,9 +1,9 @@
 import { useSelector } from 'react-redux';
 import { makeEventByTypeSelector } from 'store';
 import { RollCall } from 'features/rollCall/objects';
+import { LaoEventType } from 'features/events/objects/LaoEvent';
 import { Hash } from '../Hash';
 import { generateToken } from './Token';
-import { LaoEventType } from '../LaoEvent';
 import { PopToken } from '../PopToken';
 
 /**
