+++ fe1-web/navigation/bars/organizer/OrganizerNavigation.tsx
@@ -3,8 +3,8 @@ import { createStackNavigator } from '@react-navigation/stack';
 
 import STRINGS from 'res/strings';
 
-import Organizer from 'parts/lao/organizer/Organizer';
-import CreateEvent from 'parts/lao/organizer/eventCreation/CreateEvent';
+import Organizer from 'features/events/screens/Organizer';
+import CreateEvent from 'features/events/screens/CreateEvent';
 import WitnessScanning from 'components/WitnessScanning';
 import CreateMeeting from 'features/meeting/screens/CreateMeeting';
 import CreateRollCall from 'features/rollCall/screens/CreateRollCall';
