+++ fe1-web/store/reducers/RootReducer.ts
@@ -2,9 +2,9 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 import { persistCombineReducers } from 'redux-persist';
 
 import socialReducer from 'features/social/reducer/SocialReducer';
+import eventsReducer from 'features/events/reducer/EventsReducer';
 import laoReducer from './LaoReducer';
 import keyPairReducer from './KeyPairReducer';
-import eventsReducer from './EventsReducer';
 import messageReducer from './MessageReducer';
 import walletReducer from './WalletReducer';
