+++ fe1-web/store/reducers/index.ts
@@ -19,9 +19,7 @@ import { rootReducer } from './RootReducer';
 
 export default rootReducer;
 
-export * from './EventsReducer';
 export * from './LaoReducer';
 export * from './KeyPairReducer';
 export * from './MessageReducer';
 export * from './WalletReducer';
-export * from 'features/social/reducer/SocialReducer';
