+++ fe1-web/store/stores/index.ts
@@ -12,4 +12,3 @@
 export * from './KeyPairStore';
 export * from './OpenedLaoStore';
 export * from './WalletStore';
-export * from './EventStore';
